Made logo container sticky
This is to prevent the logo form disappearing at the top when the
height become small. Now the top end of the logo container will stick to
the top of the surrounding container (which here is at the same place
as the body to -- because the flex container is positioned absolute and
fills the whole body).