Scrolling to project detail header after showing
Before, the click on the project overview would show the project detail
card and hide the projects overview. The click might happen when the top
of the project overview is not visible anymore. This postion would also
be kept after the detail card is shown. That would mean that the user
does not see the header of the card that was just opened.

To fix this, a scroll to the project detail element is added after it
is shown. Thus, the top of the card is now focused when the overview
card is clicked.