Project overview cards layout changed
Before, all three cards where displayed in one row. That was ok, but
lead to very slim card on narrow displays.

Now the layout is changed that the first row only contains two cards
and the second row only one card.

The width has to be controlled manual as of now, by defining the column
width for the different rows.